Fix logic bug in test_speed
There was an extra "exec" call whose time is not printed,
but is included in the total time printed,
but then helper()'s return value excludes the time of the 4th exec.

This changes the results somewhat dramatically,
because the exec that is not printed is the slow one
that hydrates all the code objects.